% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{concordances}
\alias{concordances}
\title{Concordances}
\usage{
concordances(source, target)
}
\arguments{
\item{source}{Taxonomy of input vector.}

\item{target}{Taxonomy of output vector.}
}
\value{
data.table with \code{source} and \code{target} columns.
}
\description{
Get the concordances between a source taxonomy and a target taxonomy.
}
