% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{crosswalk}
\alias{crosswalk}
\title{Crosswalk}
\usage{
crosswalk(dat, source, target)
}
\arguments{
\item{dat}{Input data, with at least a \code{job} column and a \code{count} column. Can be of type data.frame or data.table.}

\item{source}{Taxonomy of input vector.}

\item{target}{Taxonomy of output vector.}
}
\value{
data.table with \code{job} column converted to target taxonomy and \code{count} column.
}
\description{
Go from a source taxonomy to a target taxonomy.
}
