% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esco.R
\docType{data}
\name{esco}
\alias{esco}
\title{ESCO occupations taxonomy}
\format{
A data frame with 2942 rows and 10 columns.
\describe{
\item{esco_label}{Label of ESCO occupation}
\item{esco_key}{Key of ESCO occupation}
\item{isco_4_label}{Label of 4-digit ISCO occupation}
\item{isco_4_key}{Key of 4-digit ISCO occupation}
\item{isco_3_label}{Label of 3-digit ISCO occupation}
\item{isco_3_key}{Key of 3-digit ISCO occupation}
\item{isco_2_label}{Label of 2-digit ISCO occupation}
\item{isco_2_key}{Key of 2-digit ISCO occupation}
\item{isco_1_label}{Label of 1-digit ISCO occupation}
\item{isco_1_key}{Key of 1-digit ISCO occupation}
}
}
\source{
\url{https://ec.europa.eu/esco/portal/occupation}
}
\usage{
esco
}
\description{
A dataset containing they hierarchy of ESCO/ISCO, with both labels and keys included.
}
\keyword{datasets}
