% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foundation_skills.R
\docType{data}
\name{foundation_skills}
\alias{foundation_skills}
\title{Foundation Skills}
\format{
A data frame with three columns, encoded in ISCO taxonomy.
\describe{
\item{preferredLabel}{ISCO official label of occupation}
\item{Occupations}{Display label}
\item{Skill}{Foundation skills}
\item{Value}{Percentage of jobs for which foundation skills are importan}
}
}
\source{
\url{https://www.cedefop.europa.eu/en/tools/skills-intelligence}
}
\usage{
foundation_skills
}
\description{
Example dataset for indicator given in ISCO classification.
}
\keyword{datasets}
