% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_soc_code.R
\name{get_soc_code}
\alias{get_soc_code}
\title{Get SOC code from label}
\usage{
get_soc_code(data, lvl = "soc_3")
}
\arguments{
\item{data, }{data.frame or data.table with two columns \code{job} and \code{value}}

\item{lvl, }{string that can take values from \code{soc_1} up to \code{soc_4}}
}
\value{
data.frame of input data with one extra column named as \code{code}
}
\description{
Adds SOC code for a particular job title.
}
