% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isco.R
\docType{data}
\name{isco}
\alias{isco}
\title{ISCO occupations taxonomy}
\format{
A data frame with two columns, encoding the ISCO taxonomy.
\describe{
\item{code}{Code of ISCO occupation. Number of digits indicate the level}
\item{preferredLabel}{Preferred label of the ISCO occupation}
}
}
\source{
\url{https://esco.ec.europa.eu}
}
\usage{
isco
}
\description{
A dataset containing the hierarchy of ISCO, with both code and
preferred label included.
}
\keyword{datasets}
