% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isco08_soc10.R
\docType{data}
\name{isco08_soc10}
\alias{isco08_soc10}
\title{ISCO_08 to SOC_10 crosswalks}
\format{
A table with the following colums.
\describe{
\item{isco08}{Codes of ISCO 4th level}
\item{soc10}{Codes of SOC detailed occupations}
\item{isco_label}{Labels of ISCO 4th level}
\item{soc_label}{Labels of SOC detailed occupations}
}
}
\source{
\url{https://ibs.org.pl}
}
\usage{
isco08_soc10
}
\description{
A table containing the "one to many" crosswalks from ISCO version 08 to
SOC10 for the 3rd hierarchical level and detailed occupations respectively.
}
\keyword{datasets}
