% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isco_concordances.R
\name{isco_concordances}
\alias{isco_concordances}
\title{ISCO Concordances}
\usage{
isco_concordances(source, target)
}
\arguments{
\item{source}{Taxonomy of input vector.}

\item{target}{Taxonomy of output vector.}
}
\value{
data.table with \code{source} and \code{target} columns.
}
\description{
Get the concordances between a source taxonomy and a target taxonomy.
}
