% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isco_crosswalk.R
\name{isco_crosswalk}
\alias{isco_crosswalk}
\title{ISCO Crosswalks}
\usage{
isco_crosswalk(dat, source, target, aggr = TRUE)
}
\arguments{
\item{dat}{Input data, with at least a \code{job} column and a \code{value} column. Can be of type \code{data.frame} or \code{data.table}.}

\item{source}{Taxonomy of input vector.}

\item{target}{Taxonomy of output vector.}

\item{aggr}{Whether or not to aggregate results by \code{target} taxonomy. Defaults to \code{TRUE}.}
}
\value{
Table of type \code{data.table} with \code{value} column converted from \code{source} taxonomy to \code{target} taxonomy.
}
\description{
Go from an ISCO hierarchical level to target higher level.
}
