% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soc10_isco08.R
\docType{data}
\name{soc10_isco08}
\alias{soc10_isco08}
\title{SOC_10 to ISCO_08 crosswalks}
\format{
A table with the following colums.
\describe{
\item{isco08}{Codes of ISCO 4th level}
\item{soc10}{Codes of SOC detailed occupations}
\item{isco_label}{Labels of ISCO 4th level}
\item{soc_label}{Labels of SOC detailed occupations}
}
}
\source{
\url{https://ibs.org.pl/en}
}
\usage{
soc10_isco08
}
\description{
A table containing the "one to many" crosswalks from SOC version 10 to
ISCO version 08 for Detailed occupations and 4th hierarchical level
respectively.
}
\keyword{datasets}
