% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soc_groups.R
\docType{data}
\name{soc_groups}
\alias{soc_groups}
\title{SOC groups}
\format{
A table with the following columns.
\describe{
\item{soc_1}{Major groups. The first two digits indicate the soc group}
\item{soc_2}{Minor groups. The first 4 digits indicate the soc group}
\item{soc_3}{Broad occupations. The first five digits indicate the soc group}
\item{soc_4}{Detailed occupations. Digits that indicate the soc group}
}
}
\source{
\url{https://ec.europa.eu/esco/portal/occupation}
}
\usage{
soc_groups
}
\description{
A table containing the hierarchy of SOC, with both labels and keys included.
}
\keyword{datasets}
