% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{crosswalk}
\alias{crosswalk}
\title{Crosswalk}
\usage{
crosswalk(dat, source, target, aggr = TRUE)
}
\arguments{
\item{dat}{Input data, with at least a \code{job} column and a \code{count} column. Can be of type \code{data.frame} or \code{data.table}.}

\item{source}{Taxonomy of input vector.}

\item{target}{Taxonomy of output vector.}

\item{aggr}{Whether or not to aggregate results by \code{target} taxonomy. Defaults to \code{TRUE}.}
}
\value{
Table of type \code{data.table} with \code{count} values converted from \code{source} taxonomy to \code{target} taxonomy.
}
\description{
Go from a source taxonomy to a target taxonomy.
}
