% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_isco_code.R
\name{get_isco_code}
\alias{get_isco_code}
\title{Get ISCO code from label}
\usage{
get_isco_code(data, lvl = 3)
}
\arguments{
\item{data, }{data.frame or data.table with two columns \code{job} and \code{value}}

\item{lvl, }{numeric value indicating the ISCO taxonomy}
}
\description{
Adds ISCO code for a particular job title
}
