% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_soc_code.R
\name{get_soc_code}
\alias{get_soc_code}
\title{Get SOC code from label}
\usage{
get_soc_code(data, lvl = "soc_3")
}
\arguments{
\item{data}{data.table with mandatory column \code{job}.}

\item{lvl}{Character string taking values between \code{soc_1} and \code{soc_4}.}
}
\value{
data.frame of input data with one extra column named as \code{code}
}
\description{
Adds column of SOC code for a particular job title. Job titles should be
given with the labels defined by SOC.
}
